Working login/authenticate mechanism. For now it's tested with the adduser/setup
Use: paster setup-app development.ini
Then: paster serve --reload development.ini

Then, run, from scripts/
./sflvault adduser test1
./sflvault setup test1 unused_url

It will generate an ElGamal key, send it to the server, server will accept.
Then the login is initiated, server accepts, sends a cryptok. Client decrypt
and resends the 'authenticate' with decrypted token (using the privkey).

Server says Authentication successful, with an authtok (to be used in further
xmlrpc requests, where authtok is required).

TODO: - mplement the authtok saving in the user session, server-side (some
tweaks to beaker are required, to recognize the authtok instead of some
cookies).
      - In the client, implement the saving and loading of configuration, and
        privkey encryption/storing/asking for password.

Asking for password is delicate, ssh connects directly to the tty for that.
Can we do the same ?