First move to OBJECTIFY the server-side of the Vault.
NOTE: grant doesn't work in this version. I copied only the contents from
xmlrpc to lib.vault. To keep the patches moving.

Had to take the __call__ function into xmlrpc.py to avoid recursive includes.

lib.vault required to include lib.base but lib.base needed lib.vault to initialise
the SFLvaultAccess object, but only controllers.xmlrpc needs it, so it's been
moved there.

now controllers.xmlrpc require lib.base
and lib.vault requires lib.base

This is all going to change when everything is moved to lib.vault (the object).

When this is done, we could use the client to communicate directly with the
vault (when it's on the same machine) with no restrictions.