Remove with_groups parameter
This parameter is only used to determine if the client will display the
list of groups. However, it is checked everywhere in the vault. To make
the code easier to read, with_groups is removed on the server side,
and verbose is used to do the filtering is done on the client side.

It is kept in the XML-RPC interfaces for compatibility, but a deprecation
notice has been added in the documentation.