Remove user from the vault
User is currently set directly on the vault, as myself_id and
myself_username. Instead of doing this, retrieve the user using
the authtok of the request and pass it to the vault function.

myself_id and myself_username is mostly used by logging functions.
This commit remove log_i, log_e and log_w and replace all of its
occurences with log_command when necessary.